backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.2
python-dotenv==1.0.0
google-generativeai==0.3.0
google-cloud-firestore==2.14.0
feedparser==6.0.10
aiohttp==3.9.1
PyJWT==2.8.1
httpx==0.25.2
